oooouhguhouhgoughg

+++ src/NeighbourTD.cpp
@@ -52,11 +52,11 @@ IntVector TranslateParticleToNeighbour(IntVector pos, int tile_size) {
 
 IntVector VecDifference(IntVector vec1, IntVector vec2) {
     int dx, dy;
-    dx = (vec1.x - vec2.x);
-    dx = (abs(dx) /*- 1*/) * NTDsignumInverted(dx);
+    dx = ((int)vec1.x - (int)vec2.x);
+    dx = (abs(dx)) * NTDsignumInverted(dx);
 
-    dy = (vec1.y - vec2.y);
-    dy = (abs(dy) /*- 1*/) * NTDsignumInverted(dy);
+    dy = ((int)vec1.y - (int)vec2.y);
+    dy = (abs(dy) ) * NTDsignumInverted(dy);
 
     return { dx, dy };
 }
